example/requirements.txt
require.py==2.0.1
tqdm==4.10.0
python-chess==0.16.1
